:ambulance: Fixed the gitignore file
close #26
