+++ .travis.yml
@@ -2,6 +2,7 @@ language: python
 python:
   - 3.7
 install:
-  - pip install -r requirements-dev.txt
+  - pip install -q pipenv
+  - pipenv install --dev
 script:
   - flake8
