+++ MANIFEST.in
@@ -1,3 +1,2 @@
 include README.md
 include LICENSE
-include pycutter/.gitignore
