+++ README.md
@@ -1,4 +1,6 @@
 # pycutter
 Cookiecutter to start projects in Python
 
+[![Python 3](https://pyup.io/repos/github/alvesgabriel/pycutter/python-3-shield.svg)](https://pyup.io/repos/github/alvesgabriel/pycutter/)
 [![Build Status](https://travis-ci.org/alvesgabriel/pycutter.svg?branch=master)](https://travis-ci.org/alvesgabriel/pycutter)
+[![Updates](https://pyup.io/repos/github/alvesgabriel/pycutter/shield.svg)](https://pyup.io/repos/github/alvesgabriel/pycutter/)
