+++ pycutter/command/gitignore.py
@@ -1,4 +1,4 @@
-# Byte-compiled / optimized / DLL files
+GITIGNORE = """# Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
 *$py.class
@@ -127,3 +127,4 @@ dmypy.json
 
 # Pyre type checker
 .pyre/
+"""
